fix: refactoring time commands and adding different conversion method
