+++ src/commands/time/modules/timeZoneData.ts
@@ -1,0 +1,5 @@
+import { TimeZone } from "./time_conversion.js"
+
+const timeZoneData: TimeZone[] = JSON.parse("[ { \"name\": \"Hawaii Standard Time (HST)\", \"value\": { \"name\": \"HST\", \"isPositive\": false, \"hours\": 10, \"minutes\": 0 } }, { \"name\": \"Alaska Standard Time (AKST)\", \"value\": { \"name\": \"AKST\", \"isPositive\": false, \"hours\": 9, \"minutes\": 0 } }, { \"name\": \"Pacific Standard Time (PST)\", \"value\": { \"name\": \"PST\", \"isPositive\": false, \"hours\": 8, \"minutes\": 0 } }, { \"name\": \"Mountain Standard Time (MST)\", \"value\": { \"name\": \"MST\", \"isPositive\": false, \"hours\": 7, \"minutes\": 0 } }, { \"name\": \"Central Standard Time (CST)\", \"value\": { \"name\": \"CST\", \"isPositive\": false, \"hours\": 6, \"minutes\": 0 } }, { \"name\": \"Eastern Standard Time (EST)\", \"value\": { \"name\": \"EST\", \"isPositive\": false, \"hours\": 5, \"minutes\": 0 } }, { \"name\": \"Atlantic Standard Time (AST)\", \"value\": { \"name\": \"AST\", \"isPositive\": false, \"hours\": 4, \"minutes\": 0 } }, { \"name\": \"Argentina Time (ART)\", \"value\": { \"name\": \"ART\", \"isPositive\": false, \"hours\": 3, \"minutes\": 0 } }, { \"name\": \"Greenwich Mean Time (GMT)\", \"value\": { \"name\": \"GMT\", \"isPositive\": true, \"hours\": 0, \"minutes\": 0 } }, { \"name\": \"Central European Time (CET)\", \"value\": { \"name\": \"CET\", \"isPositive\": true, \"hours\": 1, \"minutes\": 0 } }, { \"name\": \"Eastern European Time (EET)\", \"value\": { \"name\": \"EET\", \"isPositive\": true, \"hours\": 2, \"minutes\": 0 } }, { \"name\": \"Moscow Standard Time (MSK)\", \"value\": { \"name\": \"MSK\", \"isPositive\": true, \"hours\": 3, \"minutes\": 0 } }, { \"name\": \"Iran Standard Time (IRST)\", \"value\": { \"name\": \"IRST\", \"isPositive\": true, \"hours\": 3, \"minutes\": 30 } }, { \"name\": \"Gulf Standard Time (GST)\", \"value\": { \"name\": \"GST\", \"isPositive\": true, \"hours\": 4, \"minutes\": 0 } }, { \"name\": \"Pakistan Standard Time (PKT)\", \"value\": { \"name\": \"PKT\", \"isPositive\": true, \"hours\": 5, \"minutes\": 0 } }, { \"name\": \"Indian Standard Time (IST)\", \"value\": { \"name\": \"IST\", \"isPositive\": true, \"hours\": 5, \"minutes\": 30 } }, { \"name\": \"Bangladesh Standard Time (BST)\", \"value\": { \"name\": \"BST\", \"isPositive\": true, \"hours\": 6, \"minutes\": 0 } }, { \"name\": \"Myanmar Time (MMT)\", \"value\": { \"name\": \"MMT\", \"isPositive\": true, \"hours\": 6, \"minutes\": 30 } }, { \"name\": \"Indochina Time (ICT)\", \"value\": { \"name\": \"ICT\", \"isPositive\": true, \"hours\": 7, \"minutes\": 0 } }, { \"name\": \"China Standard Time (CNST)\", \"value\": { \"name\": \"CNST\", \"isPositive\": true, \"hours\": 8, \"minutes\": 0 } }, { \"name\": \"Japan Standard Time (JST)\", \"value\": { \"name\": \"JST\", \"isPositive\": true, \"hours\": 9, \"minutes\": 0 } }, { \"name\": \"Australian Eastern Standard Time (AEST)\", \"value\": { \"name\": \"AEST\", \"isPositive\": true, \"hours\": 10, \"minutes\": 0 } }, { \"name\": \"New Zealand Standard Time (NZST)\", \"value\": { \"name\": \"NZST\", \"isPositive\": true, \"hours\": 12, \"minutes\": 0 } }, { \"name\": \"Line Islands Time (LINT)\", \"value\": { \"name\": \"LINT\", \"isPositive\": true, \"hours\": 14, \"minutes\": 0 } } ]")
+
+export default timeZoneData
